ipa: libipa: fixedpoint: Remove unsigned workaround
Now that we no longer attempt or allow storing signed negative floats
as an unsigned integer representation, remove the workaround which
required an intermediate cast when converting floats to fixed point values.

+++ src/ipa/libipa/fixedpoint.h
@@ -69,12 +69,7 @@ public:
 	{
 		v = std::clamp(v, min, max);
 
-		/*
-		 * The intermediate cast to int is needed on arm platforms to
-		 * properly cast negative values. See
-		 * https://embeddeduse.com/2013/08/25/casting-a-negative-float-to-an-unsigned-int/
-		 */
-		return static_cast<T>(static_cast<int>(std::round(v * (1 << F)))) & bitMask;
+		return static_cast<T>(std::round(v * (1 << F))) & bitMask;
 	}
 };
